refactor(ledger): emit invoked_by and agent_name, move detection into ledger
