+++ internal/ledger/segmentledger/provider.go
@@ -39,8 +39,13 @@ func (p *provider) Track(_ context.Context, event domain.Event, properties domai
 	properties = properties.
 		Set("os", runtime.GOOS).
 		Set("arch", runtime.GOARCH).
-		Set("foundry_version", version.Info.Version()).
-		WithAIAgent(domain.NewAIAgent(os.Getenv))
+		Set("foundry_version", version.Info.Version())
+
+	if agent := ledger.AgentName(os.Getenv); agent != "" {
+		properties = properties.Set("invoked_by", "agent").Set("agent_name", agent)
+	} else {
+		properties = properties.Set("invoked_by", "unknown")
+	}
 
 	props := segment.NewProperties()
 	for k, v := range properties.Map() {
